refactor(vault-ingress): give the catalog duration reader one owner too
`expected_duration_seconds` was copied into both the auditor and preflight.
The two copies agree today, which is exactly where the catalog-date predicate
started before they drifted apart. Behavior is unchanged; the second copy is
gone.

Refs #333
